+++ composer/src/main/java/io/piano/android/composer/DependenciesProvider.kt
@@ -27,7 +27,7 @@ internal class DependenciesProvider private constructor(
         .addInterceptor(UserAgentInterceptor(userAgent))
         .addInterceptor(
             HttpLoggingInterceptor().setLevel(
-                if (BuildConfig.DEBUG)
+                if (BuildConfig.DEBUG || isLogHttpSet())
                     HttpLoggingInterceptor.Level.BODY
                 else HttpLoggingInterceptor.Level.NONE
             )