+++ composer/src/main/java/io/piano/android/composer/model/DelayBy.kt
@@ -8,6 +8,7 @@ class DelayBy(
     @JvmField val type: DelayType?,
     @JvmField val value: Int
 ) {
+    @JsonClass(generateAdapter = false)
     enum class DelayType {
         @Json(name = "time") TIME,
         @Json(name = "scroll") SCROLL