+++ plugins/src/main/kotlin/io/piano/android/dependencies/Deps.kt
@@ -1,35 +1,19 @@
-import org.gradle.api.JavaVersion
+package io.piano.android.dependencies
 
-object Config {
-    // Android config
-    const val androidBuildTools = "30.0.2"
-    const val androidMinSdk = 16
-    const val androidTargetSdk = 30
-    const val androidCompileSdk = 30
-    val compileSourceVersion = JavaVersion.VERSION_1_8
-    val compileTargetVersion = JavaVersion.VERSION_1_8
-}
-
-object Versions {
+private object Versions {
     const val kotlin = "1.4.32"
-    //Plugins
-    const val versionsPlugin = "0.38.0"
-    const val androidToolsPlugin = "4.1.3"
-    const val ktlintPlugin = "10.0.0"
-    const val dokkaPlugin = "1.4.30"
-    const val publishPlugin = "0.14.2"
 
     // AndroidX libraries
-    const val compatLibrary = "1.2.0"
-    const val androidxActivity = "1.2.2"
+    const val compatLibrary = "1.3.0"
+    const val androidxActivity = "1.2.3"
     const val annotationsLibrary = "1.2.0"
-    const val fragmentLibrary = "1.3.2"
-    const val materialLibrary = "1.3.0"
+    const val fragmentLibrary = "1.3.5"
+    const val materialLibrary = "1.4.0"
     const val lifecycleLibrary = "2.3.1"
     const val prefsLibrary = "1.1.1"
 
     // Third party Libraries
-    const val facebookLogin = "9.1.0"
+    const val facebookLogin = "11.1.0"
     const val googlePlayServices = "19.0.0"
     const val retrofit = "2.6.4"
     const val okhttp = "3.12.13"
@@ -41,21 +25,8 @@ object Versions {
     const val androidxTestCore = "1.3.0"
     const val androidxTestExtJunit = "1.1.2"
     const val mockitoKotlin = "2.2.0"
-    const val mockitoCore = "3.0.0"
+    const val mockitoCore = "3.11.2"
     const val robolectric = "4.5.1"
-}
-
-object Plugins {
-    const val kotlin = "gradle-plugin"
-    const val versions = "com.github.ben-manes.versions"
-    const val androidTools = "com.android.tools.build:gradle:${Versions.androidToolsPlugin}"
-    const val androidApp = "com.android.application"
-    const val androidLibrary = "com.android.library"
-    const val kotlinAndroid = "kotlin-android"
-    const val kotlinParcelize = "kotlin-parcelize"
-    const val ktlint = "org.jlleitschuh.gradle.ktlint"
-    const val dokka = "org.jetbrains.dokka"
-    const val publish = "com.vanniktech.maven.publish"
 }
 
 object Libs {
@@ -75,7 +46,6 @@ object Libs {
     const val timber = "com.jakewharton.timber:timber:${Versions.timber}"
     const val moshi = "com.squareup.moshi:moshi:${Versions.moshi}"
     const val moshiCodegen = "com.squareup.moshi:moshi-kotlin-codegen:${Versions.moshi}"
-    const val prefs = "androidx.preference:preference:${Versions.prefsLibrary}"
     const val prefsKtx = "androidx.preference:preference-ktx:${Versions.prefsLibrary}"